requirements.txt
streamlit==1.31.0
supabase==1.2.0
python-dotenv==1.0.0
httpx==0.24.1
postgrest<0.12.0
google-cloud-translate==3.12.0
google-cloud-texttospeech==2.14.1
gtts==2.4.0
requests==2.31.0
plotly==5.18.0
pandas==2.1.4
numpy==1.26.3
streamlit-chat==0.1.1
streamlit-option-menu==0.3.12
Pillow==10.1.0
sounddevice==0.4.6
retry-requests==2.0.0
matplotlib==3.8.2
librosa==0.10.1
python-jose==3.3.0
bcrypt==4.2.1
pydantic==2.5.2
langchain==0.0.348
openai==1.6.1
tiktoken==0.5.2
pytest==7.4.3
pytest-cov==4.1.0
azure-cognitiveservices-speech==1.33.0
